requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
# Pillow==8.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis
celery==4.4.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==2.2.20  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
# django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Added
django-filter==2.4.0
djangorestframework-jwt
simplejson==3.17.0
drf-jwt-2fa==0.3.0
django-templated-mail==1.1.1
twilio


# Channels
channels==2.4.0
channels_redis==3.1.0
websocket_client==0.57.0
asgiref==3.3.4

xlrd==1.2.0
django-storages[boto3]==1.7.2  # https://github.com/jschneier/django-storages
boto3==1.17.48
botocore==1.20.48
Collectfast==0.5.2
django-searchable-encrypted-fields==0.1.4

daphne==2.5.0 #https://github.com/django/daphne

bitarray==1.2.2

django-celery-results == 2.0.1
django-celery-email==3.0.0

# Reporting
django-weasyprint == 1.0.1
PyPDF2==1.28.1

#Request
requests==2.22.0
cryptography == 3.4.7
cairosvg == 2.5.2
autobahn == 21.3.1

# Screenshot
html2image==2.0.1


graphene-django
